# docs/api/openapi.yaml
openapi: 3.1.0
info:
  title: Nexus Agent Model Hub API
  version: 0.1.0
  summary: Multi-tenant auth and Agent tool-calling API
servers:
  - url: http://localhost:4000/api/v1
tags:
  - name: health
  - name: auth
  - name: tenants
  - name: users
  - name: agents
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
    apiKeyAuth:
      type: apiKey
      in: header
      name: X-API-Key
  schemas:
    ErrorResponse:
      type: object
      required: [error, message]
      properties:
        error:
          type: string
        message:
          type: string
        details:
          type: object
          additionalProperties: true
    TokenPair:
      type: object
      required: [accessToken, refreshToken, expiresIn]
      properties:
        accessToken:
          type: string
        refreshToken:
          type: string
        expiresIn:
          type: string
    Tenant:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        slug:
          type: string
        description:
          type: string
          nullable: true
        plan:
          type: string
        isActive:
          type: boolean
    User:
      type: object
      properties:
        id:
          type: string
        email:
          type: string
          format: email
        displayName:
          type: string
        locale:
          type: string
        role:
          type: string
    AgentDefinition:
      type: object
      properties:
        type:
          type: string
        name:
          type: string
        description:
          type: string
        version:
          type: string
        inputSchema:
          type: object
          additionalProperties: true
        outputSchema:
          type: object
          additionalProperties: true
paths:
  /health:
    get:
      tags: [health]
      operationId: getHealth
      responses:
        "200":
          description: Healthy
        "503":
          description: Degraded
  /auth/register:
    post:
      tags: [auth]
      operationId: registerTenantOwner
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [tenantSlug, tenantName, email, password, displayName]
              properties:
                tenantSlug:
                  type: string
                tenantName:
                  type: string
                email:
                  type: string
                  format: email
                password:
                  type: string
                displayName:
                  type: string
                locale:
                  type: string
                  enum: [zh-CN, en-US]
      responses:
        "201":
          description: Registered
        "409":
          description: Tenant already exists
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /auth/login:
    post:
      tags: [auth]
      operationId: login
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [tenantSlug, email, password]
              properties:
                tenantSlug:
                  type: string
                email:
                  type: string
                  format: email
                password:
                  type: string
                otpCode:
                  type: string
      responses:
        "200":
          description: Login success
        "401":
          description: Invalid credentials
  /auth/refresh:
    post:
      tags: [auth]
      operationId: refreshSession
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [refreshToken]
              properties:
                refreshToken:
                  type: string
      responses:
        "200":
          description: New token pair
  /auth/me:
    get:
      tags: [auth]
      operationId: getCurrentUser
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Current user
  /tenants/current:
    get:
      tags: [tenants]
      operationId: getCurrentTenant
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Current tenant
    patch:
      tags: [tenants]
      operationId: updateCurrentTenant
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Updated tenant
  /tenants/current/audit-logs:
    get:
      tags: [tenants]
      operationId: listAuditLogs
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Tenant audit logs
  /users:
    get:
      tags: [users]
      operationId: listUsers
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Tenant users
    post:
      tags: [users]
      operationId: createUser
      security:
        - bearerAuth: []
      responses:
        "201":
          description: Created user
  /users/{id}:
    patch:
      tags: [users]
      operationId: updateUser
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Updated user
    delete:
      tags: [users]
      operationId: deactivateUser
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
      responses:
        "204":
          description: Deactivated
  /agents:
    get:
      tags: [agents]
      operationId: listAgents
      security:
        - bearerAuth: []
        - apiKeyAuth: []
      responses:
        "200":
          description: Registered agents
          content:
            application/json:
              schema:
                type: object
                properties:
                  agents:
                    type: array
                    items:
                      $ref: "#/components/schemas/AgentDefinition"
  /agents/call:
    post:
      tags: [agents]
      operationId: callAgent
      security:
        - bearerAuth: []
        - apiKeyAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [agentType, input]
              properties:
                agentType:
                  type: string
                input:
                  type: object
                  additionalProperties: true
                timeoutMs:
                  type: integer
      responses:
        "200":
          description: Agent completed
        "504":
          description: Agent timed out
  /agents/runs:
    get:
      tags: [agents]
      operationId: listAgentRuns
      security:
        - bearerAuth: []
        - apiKeyAuth: []
      responses:
        "200":
          description: Agent run history
  /agents/runs/{id}:
    get:
      tags: [agents]
      operationId: getAgentRun
      security:
        - bearerAuth: []
        - apiKeyAuth: []
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Single Agent run